08/03/2022 - New API for Desarrollo web entorno cliente - Modify UsuarioPDO, DepartamentoPDO on model - Modify cMtoDepartamentos, cInicioPrivado, cCambiarPassword on controller - Modify vMtoUsuarios, layaout on view - Modify CSS - Add new files for DiagramaClases - Update README, configDBPDO - Update index - Modify htaccess
