+++ README.md
@@ -5,7 +5,7 @@ Desarrollo de la aplicación final del Curso.
 
 **Fecha Inicio Proyecto: 24/01/2022**
 
-**Ultima Actualización: 24/02/2022**
+**Ultima Actualización: 08/03/2022**
 
 ## Descripción 
 Desarrollo de la aplicación final del Curso.
@@ -50,13 +50,13 @@ Desarrollo de la aplicación final del Curso.
 ## Instrucciones de despliegue
 Hola como estas, gracias por la compra o descarga de esta aplicacion, a continuacion te escribo las instrucciones de instalacion:
 
-1.- Copiar y descomprimir el archivo en la carpeta \public_html, al final tendras una carpeta llamada “207DWESAplicaccionFinalAlberto2022”, a la cual accederas desde el navegador como -> http://127.0.0.1/207DWESAplicaccionFinalAlberto2022
-2.- En caso de ser necesario modificar el archivo configDBPDO en config/configDBPDO.php, para los datos asociados al acceso a la base de datos.
-3.- Buscar los archivos CreaDB207DWESProyectoTema5.sql, CargaInicialDB207DWESProyectoTema5, BorraDB207DWESProyectoTema5.sql en 207DWESAplicaccionFinalAlberto2022/scriptDB/nombreArchivo.sql, una vez localizado copiar y pegar el contenido en una ventana de MySQL, PHPMyAdmin o MySQL WorkBench
+1. Copiar y descomprimir el archivo en la carpeta \public_html, al final tendras una carpeta llamada “207DWESAplicaccionFinalAlberto2022”, a la cual accederas desde el navegador como -> http://127.0.0.1/207DWESAplicaccionFinalAlberto2022
+2. En caso de ser necesario modificar el archivo configDBPDO en config/configDBPDO.php, para los datos asociados al acceso a la base de datos.
+3. Buscar los archivos CreaDB207DWESProyectoTema5.sql, CargaInicialDB207DWESProyectoTema5, BorraDB207DWESProyectoTema5.sql en 207DWESAplicaccionFinalAlberto2022/scriptDB/nombreArchivo.sql, una vez localizado copiar y pegar el contenido en una ventana de MySQL, PHPMyAdmin o MySQL WorkBench
 
-Login :  home496960781.1and1-data.host
-Usuario: u74942863-daw207
-Password: daw2_Sauces
+- Login :  home496960781.1and1-data.host
+- Usuario: u74942863-daw207
+- Password: daw2_Sauces
 
 Dudas y sugerencias a mi direccion email alberto.ferram@educa.jcyl.es
 Powered By https://daw207.ieslossauces.es/
