+++ controller/cCambiarPassword.php
@@ -30,7 +30,7 @@ if(isset($_REQUEST['aceptar'])){
     $aErrores['passwordNueva'] = validacionFormularios::validarPassword($_REQUEST['PasswordNueva'], 8, 1, 1, OBLIGATORIO); //Compruebo si la password esta bien rellenada
     $aErrores['passwordRepetir'] = validacionFormularios::validarPassword($_REQUEST['RepetirPasswordNueva'], 8, 1, 1, OBLIGATORIO); //Compruebo si la password repetida esta bien rellenada)
     
-    if(!UsuarioPDO::validarUsuario($_SESSION['usuario207DWESAplicaccionFinal']->getCodUsuario(), $_REQUEST['PasswordActual'])){
+    if(!UsuarioPDO::validarUsuario($_SESSION['usuario207DWESAplicaccionFinalAlberto']->getCodUsuario(), $_REQUEST['PasswordActual'])){
         $aErrores['passwordActual'] = 'Password incorrecta';
         $entradaOK = false;//Le doy el valor false a entradaOK
     }else{
@@ -52,7 +52,7 @@ if(isset($_REQUEST['aceptar'])){
 }
 
 if($entradaOK){
-    $_SESSION['usuario207DWESAplicaccionFinal'] = UsuarioPDO::cambiarPassword($_SESSION['usuario207DWESAplicaccionFinal'], $_REQUEST['PasswordNueva']);
+    $_SESSION['usuario207DWESAplicaccionFinalAlberto'] = UsuarioPDO::cambiarPassword($_SESSION['usuario207DWESAplicaccionFinalAlberto'], $_REQUEST['PasswordNueva']);
     $_SESSION['paginaEnCurso'] = 'micuenta'; //Asigno a la pagina en curso la pagina micuenta
     header('Location: index.php'); //Redireciono de nuevo a micuenta
     exit;
