+++ view/layout.php
@@ -94,7 +94,7 @@
                     <img src="../207DWESAplicaccionFinalAlberto2022/webroot/css/img/github.png" class="imagegithub" alt="IconoGitHub" title="Repositorio GitHub"/>
                 </a>
             </div>
-            <p><a class="miweb" href="http://daw207.ieslossauces.es/index.php">Alberto Fernández Ramírez</a><a>&nbsp;&copy;</a> 2022 &nbsp;&nbsp;Release 4.0 &nbsp;&nbsp;Ultima actualización: 24/02/2022</p>
+            <p><a class="miweb" href="http://daw207.ieslossauces.es/index.php">Alberto Fernández Ramírez</a><a>&nbsp;&copy;</a> 2022 &nbsp;&nbsp;Release 4.1 &nbsp;&nbsp;Ultima actualización: 08/03/2022</p>
         </footer>
     </body>
 </html>
