+++ webroot/css/estiloejercicio.css
@@ -1669,7 +1669,7 @@ td{
 .fieldsetusuarios{
     border-radius: 32px;
     width: auto;
-    height: 370px;
+    height: 900px;
     margin: auto;
     margin-left: 50px;
     margin-right: 50px;
